with this alignment query remove finder_image from Pe
